requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
numpy-financial>=1.0.0
plotly>=5.18.0
requests>=2.31.0
scipy>=1.12.0
gridstatusio>=0.3.0
fpdf2>=2.7.0
matplotlib>=3.8.0
# Control service dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.26.0
paho-mqtt>=2.0.0
pydantic-settings>=2.1.0
